refs #171 -- move modules around: splitted the `usa.py` into multiple files
