+++ workalendar/usa/hawaii.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from .core import UnitedStates
+
+
+class Hawaii(UnitedStates):
+    """Hawaii"""
+    pass
