+++ workalendar/usa/oregon.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from .core import UnitedStates, FloatToNearestWeekdayMixin
+
+
+class Oregon(UnitedStates, FloatToNearestWeekdayMixin):
+    """Oregon"""
+
+    def get_variable_days(self, year):
+        days = super(Oregon, self).get_variable_days(year)
+        days = self.float(days)
+        return days
+
+    def get_fixed_holidays(self, year):
+        days = super(Oregon, self).get_fixed_holidays(year)
+        days = self.float(days, year)
+        return days
