+++ workalendar/usa/utah.py
@@ -1,0 +1,22 @@
+# -*- coding: utf-8 -*-
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from .core import UnitedStates, FloatToNearestWeekdayMixin
+
+
+class Utah(UnitedStates, FloatToNearestWeekdayMixin):
+    """Utah"""
+    FIXED_HOLIDAYS = UnitedStates.FIXED_HOLIDAYS + (
+        (7, 24, "Pioneer Day"),
+    )
+
+    def get_variable_days(self, year):
+        days = super(Utah, self).get_variable_days(year)
+        days = self.float(days)
+        return days
+
+    def get_fixed_holidays(self, year):
+        days = super(Utah, self).get_fixed_holidays(year)
+        days = self.float(days, year)
+        return days
