+++ workalendar/usa/wyoming.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+from .core import UnitedStates, FloatToNearestWeekdayMixin
+
+
+class Wyoming(UnitedStates, FloatToNearestWeekdayMixin):
+    """Wyoming"""
+    def get_variable_days(self, year):
+        days = super(Wyoming, self).get_variable_days(year)
+        days = self.float(days)
+        return days
+
+    def get_fixed_holidays(self, year):
+        days = super(Wyoming, self).get_fixed_holidays(year)
+        days = self.float(days, year)
+        return days
